refactor(structure): Refactor directory structure
